Draft changes to convert hp/max HP to 16bit
Issue #43
-e 
Original/build commit ID: 807b241a48bcdf9087a6404183a819ff6b520635

+++ game_state.h
@@ -90,10 +90,10 @@ typedef struct {
     UINT8 hour;
 
     // Max HP the user can have
-    UINT8 max_hp;
+    UINT16 max_hp;
 
     // Current amount of HP the user has
-    UINT8 hp;
+    UINT16 hp;
 
     // Current amount of strength, intelligence and charm
     unsigned int intelligence;
